+++ trace_simexp/info_file/execute.py
@@ -22,9 +22,8 @@ def read(exec_info_contents: list) -> tuple:
     for num_line, line in enumerate(exec_info_contents):
 
         # The fullname of pre-process info file
-        if "prepro.info Fullname" in line:
+        if "prepro.info File" in line:
             prepro_info = line.split("-> ")[-1].strip()
-
         # The base directory name
         if "Base Directory Name" in line:
             base_dir = line.split("-> ")[-1].strip()
@@ -63,8 +62,9 @@ def write(inputs: dict):
     :return: the exec.info file with the specified filename
     """
     from datetime import datetime
+    from . import common
 
-    header = ["prepro.info Name", "prepro.info Fullname",
+    header = ["prepro.info Name", "prepro.info File",
               "Base Directory Name", "Base Case Name",
               "List of Parameters Name", "Design Matrix Name",
               "TRACE Executable", "XTV2DMX Executable",
@@ -128,21 +128,6 @@ def write(inputs: dict):
 
         # Samples to Run
         info_file.writelines("{:<30s}{:3s}\n" .format(header[10], "->"))
-
-        for i in range(int(len(inputs["samples"])/10)):
-            offset1 = i*10
-            offset2 = (i+1)*10
-            for j in range(offset1, offset2 - 1):
-                info_file.writelines(" {:5d} " .format(inputs["samples"][j]))
-            info_file.writelines(" {:5d}\n"
-                                 .format(inputs["samples"][offset2-1]))
-
-        offset1 = int(len(inputs["samples"])/10) * 10
-        offset2 = len(inputs["samples"])
-        if offset2 > offset1:
-            for i in range(offset1, offset2):
-                info_file.writelines(" {:5d} " .format(inputs["samples"][i]))
-            info_file.writelines("\n")
-
+        common.write_by_tens(inputs["samples"], "5d", info_file)
         # Mark the end of samples
         info_file.writelines("***  End of Samples  ***\n")